backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.16
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.1.1
flake8-isort==6.1.1
gunicorn==23.0.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
